% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFiguresAndTables.r
\name{addFiguresAndTables}
\alias{addFiguresAndTables}
\title{Function addFiguresAndTables}
\usage{
addFiguresAndTables(text, captions = NA, captions2 = NA,
  format = "pdf")
}
\arguments{
\item{text}{Character element.}
}
\description{
Replaces notions like "insert table" or "insert figure" with a blanco LaTeX-figure (fig) oder -table (tab).
}
\examples{
addFiguresAndTable("please insert Figure 1 about here")
}
\keyword{helper}
