% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLettrine.r
\name{addLettrine}
\alias{addLettrine}
\title{Function addLettrine}
\usage{
addLettrine(text)
}
\arguments{
\item{text}{Character element.}
}
\description{
Replaces the first character with a LaTeX-letttrine.
}
\examples{
lettrine("Hello world")
}
\keyword{helper}
