% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSections.r
\name{addSections}
\alias{addSections}
\title{Function addSections}
\usage{
addSections(text)
}
\arguments{
\item{text}{Character element.}
}
\description{
Declares paragraphs with a few number of words as LaTeX-sections.
}
\examples{
sections("Hello\\nworld")
}
\keyword{helper}
