% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPassage.r
\name{extractPassage}
\alias{extractPassage}
\title{Function extractPassage}
\usage{
extractPassage(text, startReg = "\\nReferences\\n", endReg = "$")
}
\description{
Returns the substring of a text between the first match of a regular expression startReg and the first match of a regular expression endReg.
}
