% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitations.r
\name{getCitations}
\alias{getCitations}
\title{Function getCitations}
\usage{
getCitations(text)
}
\description{
Returns a data.frame containing citations formatted according to APA-style, as well as suggestions for labels and LaTeX-Code to exchange the citations with in order to add labels after each citation.
}
