% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutManuscript.r
\name{layoutManuscript}
\alias{layoutManuscript}
\title{Function layoutManuscript}
\usage{
layoutManuscript(path = "C:/Users/Andreas/Documents/MyLaTeX/texR",
  doc = "Manuscript.docx", abstractStart = "Abstract\\n",
  abstractEnd = "\\nKeywords:", keywordsStart = "\\nKeywords:",
  keywordsEnd = "INTRODUCTION: LITERATURE REVIEW",
  bodyStart = "INTRODUCTION: LITERATURE REVIEW",
  bodyEnd = "\\nElectronic Supplementary Material\\n",
  referencesStart = "\\nReferences\\n", referencesEnd = "$",
  title = "Title", runhead = "Short Title",
  authors = "Andreas Fischer, Daniel V. Holt, and Joachim Funke",
  affiliations = "Department of Psychology, Heidelberg University",
  doi = "10.11588/jddm.2018.1.57846", linenumbers = T,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  volume = (as.numeric(substr(Sys.Date(), 1, 4)) - 2014), number = 1,
  type = "Editorial", received = "\\\\customtoday\\n",
  accepted = "\\\\customtoday\\n", published = "\\\\customtoday\\n",
  supplement = NULL,
  declarations = "The authors declare they have no conflict of interests.",
  contributions = "All authors contributed equally to this paper.",
  editor = "Andreas Fischer",
  creativecommons = "This work is licensed under a Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International License.",
  citation = NA, save = T, ...)
}
\description{
Generates a tex-file compatible with JDDMLA03.cls.
}
